+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -135,8 +135,6 @@ export interface AppSettings {
 	amountOfEdgePreviews: number
 	edgeHeight: number
 	scaling: Vector3
-	camera: Vector3
-	cameraTarget: Vector3
 	hideFlatBuildings: boolean
 	invertHeight: boolean
 	invertArea: boolean