+++ visualization/app/codeCharta/state/effects/unfocusNodesOnLoadingMap/unfocusNodesOnLoadingMap.effect.ts
@@ -1,10 +1,10 @@
 import { Inject, Injectable } from "@angular/core"
 import { filter, map } from "rxjs"
-import { isAction } from "../../util/reduxHelper"
-import { createEffect } from "../angular-redux/effects/createEffect"
-import { Actions, ActionsToken } from "../angular-redux/effects/effects.module"
-import { IsLoadingFileActions, SetIsLoadingFileAction } from "../store/appSettings/isLoadingFile/isLoadingFile.actions"
-import { unfocusAllNodes } from "../store/dynamicSettings/focusedNodePath/focusedNodePath.actions"
+import { isAction } from "../../../util/reduxHelper"
+import { createEffect } from "../../angular-redux/effects/createEffect"
+import { Actions, ActionsToken } from "../../angular-redux/effects/effects.module"
+import { IsLoadingFileActions, SetIsLoadingFileAction } from "../../store/appSettings/isLoadingFile/isLoadingFile.actions"
+import { unfocusAllNodes } from "../../store/dynamicSettings/focusedNodePath/focusedNodePath.actions"
 
 @Injectable()
 export class UnfocusNodesOnLoadingMapEffect {