+++ visualization/app/codeCharta/state/store/appSettings/appSettings.reducer.ts
@@ -1,7 +1,6 @@
 import { colorLabels } from "./colorLabels/colorLabels.reducer"
 import { showMetricLabelNodeName } from "./showMetricLabelNodeName/showMetricLabelNodeName.reducer"
 import { showMetricLabelNameValue } from "./showMetricLabelNameValue/showMetricLabelNameValue.reducer"
-import { cameraTarget } from "./cameraTarget/cameraTarget.reducer"
 import { isAttributeSideBarVisible } from "./isAttributeSideBarVisible/isAttributeSideBarVisible.reducer"
 import { sortingOrderAscending } from "./sortingOrderAscending/sortingOrderAscending.reducer"
 import { isLoadingFile } from "./isLoadingFile/isLoadingFile.reducer"
@@ -13,7 +12,6 @@ import { isWhiteBackground } from "./isWhiteBackground/isWhiteBackground.reducer
 import { dynamicMargin } from "./dynamicMargin/dynamicMargin.reducer"
 import { invertHeight } from "./invertHeight/invertHeight.reducer"
 import { hideFlatBuildings } from "./hideFlatBuildings/hideFlatBuildings.reducer"
-import { camera } from "./camera/camera.reducer"
 import { scaling } from "./scaling/scaling.reducer"
 import { edgeHeight } from "./edgeHeight/edgeHeight.reducer"
 import { amountOfEdgePreviews } from "./amountOfEdgePreviews/amountOfEdgePreviews.reducer"
@@ -32,7 +30,6 @@ const appSettings = combineReducers({
 	colorLabels,
 	showMetricLabelNodeName,
 	showMetricLabelNameValue,
-	cameraTarget,
 	isAttributeSideBarVisible,
 	sortingOrderAscending,
 	isLoadingFile,
@@ -46,7 +43,6 @@ const appSettings = combineReducers({
 	invertHeight,
 	invertArea,
 	hideFlatBuildings,
-	camera,
 	scaling,
 	edgeHeight,
 	amountOfEdgePreviews,