+++ visualization/app/codeCharta/util/dataMocks.ts
@@ -1593,8 +1593,6 @@ export const STATE: State = {
 		},
 		edgeHeight: 4,
 		scaling: new Vector3(1, 1.8, 1),
-		camera: new Vector3(0, 300, 1000),
-		cameraTarget: new Vector3(177, 0, 299),
 		hideFlatBuildings: true,
 		invertHeight: true,
 		invertArea: false,
@@ -1651,8 +1649,6 @@ export const DEFAULT_STATE: State = {
 			neutral: false
 		},
 		edgeHeight: 4,
-		camera: new Vector3(0, 300, 1000),
-		cameraTarget: new Vector3(177, 0, 299),
 		dynamicMargin: true,
 		hideFlatBuildings: false,
 		invertHeight: false,
@@ -1762,8 +1758,6 @@ export const PARTIAL_SETTINGS: RecursivePartial<Settings> = {
 		amountOfEdgePreviews: 5,
 		edgeHeight: 4,
 		scaling: new Vector3(1, 1.8, 1),
-		camera: new Vector3(0, 300, 1000),
-		cameraTarget: new Vector3(1, 1, 1),
 		mapColors: DEFAULT_STATE.appSettings.mapColors
 	}
 }