+++ visualization/app/codeCharta/util/scenarioHelper.spec.ts
@@ -17,34 +17,23 @@ describe("scenarioHelper", () => {
 
 	describe("importScenarios", () => {
 		it("should convert vectors when importing scenarios", () => {
-			scenarios[0].settings.appSettings.camera = { x: 0, y: 1, z: 2 } as Vector3
-			expect(scenarios[0].settings.appSettings.camera.clone).toBeUndefined()
-
+			scenarios[0].camera = {
+				camera: { x: 0, y: 1, z: 2 } as Vector3,
+				cameraTarget: { x: 0, y: 1, z: 2 } as Vector3
+			}
 			const result = ScenarioHelper.importScenarios(scenarios)
 
-			expect(result[0].settings.appSettings.camera.clone).not.toBeUndefined()
-		})
-
-		it("should assume 0 as default value if only one dimension is given", () => {
-			scenarios[0].settings.appSettings.camera = { y: 24 } as Vector3
-			expect(scenarios[0].settings.appSettings.camera.clone).toBeUndefined()
-
-			const result = ScenarioHelper.importScenarios(scenarios)
-
-			expect(result[0].settings.appSettings.camera.clone).not.toBeUndefined()
-			expect(result[0].settings.appSettings.camera.x).toBe(1)
-			expect(result[0].settings.appSettings.camera.y).toBe(24)
-			expect(result[0].settings.appSettings.camera.z).toBe(1)
+			expect(result[0].camera.camera.clone).not.toBeUndefined()
 		})
 	})
 	describe("getScenarioSettingsByName", () => {
 		it("should return Settings for Scenario1", () => {
-			const result = ScenarioHelper.getScenarioSettingsByName("Scenario1")
+			const result = ScenarioHelper.getScenarioSettings(ScenarioHelper["scenarios"].get("Scenario1"))
 
 			expect(result).toEqual(PARTIAL_SETTINGS)
 		})
 		it("should return Settings for Scenario2", () => {
-			const result = ScenarioHelper.getScenarioSettingsByName("Scenario2")
+			const result = ScenarioHelper.getScenarioSettings(ScenarioHelper["scenarios"].get("Scenario2"))
 			const expected = {
 				dynamicSettings: { heightMetric: "mcc" },
 				appSettings: { amountOfTopLabels: 31, scaling: new Vector3(1, 1.8, 1) }
@@ -53,7 +42,7 @@ describe("scenarioHelper", () => {
 			expect(result).toEqual(expected)
 		})
 		it("should return an settingsobject with empty dynamic- and appSettings if the list does not contain the scenario", () => {
-			const result = ScenarioHelper.getScenarioSettingsByName("someScenario")
+			const result = ScenarioHelper.getScenarioSettings(undefined)
 			const expected: RecursivePartial<Settings> = { dynamicSettings: {}, appSettings: {} }
 
 			expect(result).toEqual(expected)